+++ app/decorators/case/overturned_ico/base_decorator.rb
@@ -1,0 +1,6 @@
+class Case::OverturnedICO::BaseDecorator < Case::BaseDecorator
+
+  def internal_deadline
+    I18n.l(object.internal_deadline, format: :default)
+  end
+end
