+++ app/decorators/case/overturned_ico/sar_decorator.rb
@@ -1,8 +1,4 @@
-class Case::OverturnedICO::SARDecorator < Case::BaseDecorator
-
-  def internal_deadline
-    I18n.l(object.internal_deadline, format: :default)
-  end
+class Case::OverturnedICO::SARDecorator < Case::OverturnedICO::BaseDecorator
 
   def original_case_description
     "ICO appeal (SAR) #{object.original_ico_appeal.number}"
